requirements.txt
python-decouple==3.3
Flask==1.1.1
Flask-Cors==3.0.8